requirements.txt
selenium==4.18.1
pytest==8.1.1
pytest-selenium==4.1.0
pytest-html==4.1.1
pyyaml~=6.0
requests==2.31.0
